requirements.txt
# 全自动微信公众号推文系统 - 依赖清单
requests
beautifulsoup4
selenium
webdriver-manager
markdown
icrawler
Pillow
easyocr
python-dotenv
loguru
requests-cache
rapidfuzz
